Upgrade label functions to 64-bit

+++ labels.h
@@ -12,12 +12,12 @@
 extern char lprefix[PREFIX_MAX];
 extern char lpostfix[PREFIX_MAX];
 
-bool lookup_label(char *label, int32_t *segment, int32_t *offset);
+bool lookup_label(char *label, int32_t *segment, int64_t *offset);
 bool is_extern(char *label);
-void define_label(char *label, int32_t segment, int32_t offset, char *special,
+void define_label(char *label, int32_t segment, int64_t offset, char *special,
 		  bool is_norm, bool isextrn, struct ofmt *ofmt,
                   efunc error);
-void redefine_label(char *label, int32_t segment, int32_t offset, char *special,
+void redefine_label(char *label, int32_t segment, int64_t offset, char *special,
                     bool is_norm, bool isextrn, struct ofmt *ofmt,
                     efunc error);
 void define_common(char *label, int32_t segment, int32_t size, char *special,
