+++ output/outelf64.c
@@ -218,7 +218,7 @@ static struct Symbol *lastsym;
 
 void stabs64_init(struct ofmt *, void *, FILE *, efunc);
 void stabs64_linenum(const char *filename, int32_t linenumber, int32_t);
-void stabs64_deflabel(char *, int32_t, int32_t, int, char *);
+void stabs64_deflabel(char *, int32_t, int64_t, int, char *);
 void stabs64_directive(const char *, const char *);
 void stabs64_typevalue(int32_t);
 void stabs64_output(int, void *);
@@ -456,7 +456,7 @@ static int32_t elf_section_names(char *name, int pass, int *bits)
     return sects[i]->index;
 }
 
-static void elf_deflabel(char *name, int32_t segment, int32_t offset,
+static void elf_deflabel(char *name, int32_t segment, int64_t offset,
                          int is_global, char *special)
 {
     int pos = strslen;
@@ -1389,7 +1389,7 @@ void stabs64_linenum(const char *filename, int32_t linenumber, int32_t segto)
     currentline = linenumber;
 }
 
-void stabs64_deflabel(char *name, int32_t segment, int32_t offset, int is_global,
+void stabs64_deflabel(char *name, int32_t segment, int64_t offset, int is_global,
                     char *special)
 {
     (void)name;
